Restored functionality of old dummy dakota
