+++ force_bdss/core_plugins/dummy/dummy_dakota/dakota_model.py
@@ -1,5 +1,6 @@
+from traits.api import List, String
 from force_bdss.api import BaseMCOModel
 
 
 class DummyDakotaModel(BaseMCOModel):
-    pass
+    value_types = List(String)
